Factored out PropertiesWidget from PropertiesDock
This is to allow the PropertiesWidget to be potentially also used in the
ProjectPropertiesDialog, though I'm sure there will be some additional
challenges with that.
