+++ src/tiled/propertiesdock.cpp
@@ -20,481 +20,39 @@
 
 #include "propertiesdock.h"
 
-#include "actionmanager.h"
-#include "addpropertydialog.h"
-#include "changeproperties.h"
-#include "clipboardmanager.h"
-#include "documentmanager.h"
-#include "mapdocument.h"
-#include "mapobject.h"
-#include "propertybrowser.h"
-#include "tile.h"
-#include "tileset.h"
-#include "utils.h"
-#include "variantpropertymanager.h"
+#include "propertieswidget.h"
 
-#include <QAction>
 #include <QEvent>
-#include <QInputDialog>
-#include <QKeyEvent>
-#include <QToolBar>
-#include <QUndoStack>
-#include <QVBoxLayout>
-#include <QMenu>
-#include <QFileInfo>
 
 namespace Tiled {
 
 PropertiesDock::PropertiesDock(QWidget *parent)
     : QDockWidget(parent)
-    , mDocument(nullptr)
-    , mPropertyBrowser(new PropertyBrowser)
+    , mPropertiesWidget(new PropertiesWidget(this))
 {
     setObjectName(QLatin1String("propertiesDock"));
+    setWidget(mPropertiesWidget);
 
-    mActionAddProperty = new QAction(this);
-    mActionAddProperty->setEnabled(false);
-    mActionAddProperty->setIcon(QIcon(QLatin1String(":/images/16/add.png")));
-    connect(mActionAddProperty, &QAction::triggered,
-            this, &PropertiesDock::openAddPropertyDialog);
-
-    mActionRemoveProperty = new QAction(this);
-    mActionRemoveProperty->setEnabled(false);
-    mActionRemoveProperty->setIcon(QIcon(QLatin1String(":/images/16/remove.png")));
-    mActionRemoveProperty->setShortcuts(QKeySequence::Delete);
-    connect(mActionRemoveProperty, &QAction::triggered,
-            this, &PropertiesDock::removeProperties);
-
-    mActionRenameProperty = new QAction(this);
-    mActionRenameProperty->setEnabled(false);
-    mActionRenameProperty->setIcon(QIcon(QLatin1String(":/images/16/rename.png")));
-    connect(mActionRenameProperty, &QAction::triggered,
-            this, &PropertiesDock::renameProperty);
-
-    Utils::setThemeIcon(mActionAddProperty, "add");
-    Utils::setThemeIcon(mActionRemoveProperty, "remove");
-    Utils::setThemeIcon(mActionRenameProperty, "rename");
-
-    QToolBar *toolBar = new QToolBar;
-    toolBar->setFloatable(false);
-    toolBar->setMovable(false);
-    toolBar->setIconSize(Utils::smallIconSize());
-    toolBar->addAction(mActionAddProperty);
-    toolBar->addAction(mActionRemoveProperty);
-    toolBar->addAction(mActionRenameProperty);
-
-    QWidget *widget = new QWidget(this);
-    QVBoxLayout *layout = new QVBoxLayout(widget);
-    layout->setContentsMargins(0, 0, 0, 0);
-    layout->setSpacing(0);
-    layout->addWidget(mPropertyBrowser);
-    layout->addWidget(toolBar);
-    widget->setLayout(layout);
-
-    setWidget(widget);
-
-    mPropertyBrowser->setContextMenuPolicy(Qt::CustomContextMenu);
-    connect(mPropertyBrowser, &PropertyBrowser::customContextMenuRequested,
-            this, &PropertiesDock::showContextMenu);
-    connect(mPropertyBrowser, &PropertyBrowser::selectedItemsChanged,
-            this, &PropertiesDock::updateActions);
+    connect(mPropertiesWidget, &PropertiesWidget::bringToFront,
+            this, &PropertiesDock::bringToFront);
 
     retranslateUi();
 }
 
 void PropertiesDock::setDocument(Document *document)
 {
-    if (mDocument == document)
-        return;
-
-    if (mDocument)
-        mDocument->disconnect(this);
-
-    mDocument = document;
-    mPropertyBrowser->setDocument(document);
-
-    if (document) {
-        connect(document, &Document::currentObjectChanged,
-                this, &PropertiesDock::currentObjectChanged);
-        connect(document, &Document::editCurrentObject,
-                this, &PropertiesDock::bringToFront);
-
-        connect(document, &Document::propertyAdded,
-                this, &PropertiesDock::updateActions);
-        connect(document, &Document::propertyRemoved,
-                this, &PropertiesDock::updateActions);
-
-        currentObjectChanged(document->currentObject());
-    } else {
-        currentObjectChanged(nullptr);
-    }
-}
-
-void PropertiesDock::bringToFront()
-{
-    show();
-    raise();
-    mPropertyBrowser->setFocus();
+    mPropertiesWidget->setDocument(document);
 }
 
 void PropertiesDock::selectCustomProperty(const QString &name)
 {
     bringToFront();
-    mPropertyBrowser->selectCustomProperty(name);
-}
-
-static bool anyObjectHasProperty(const QList<Object*> &objects, const QString &name)
-{
-    for (Object *obj : objects) {
-        if (obj->hasProperty(name))
-            return true;
-    }
-    return false;
-}
-
-void PropertiesDock::currentObjectChanged(Object *object)
-{
-    mPropertyBrowser->setObject(object);
-
-    bool editingTileset = mDocument && mDocument->type() == Document::TilesetDocumentType;
-    bool isTileset = object && object->isPartOfTileset();
-    bool enabled = object && (!isTileset || editingTileset);
-
-    mPropertyBrowser->setEnabled(object);
-    mActionAddProperty->setEnabled(enabled);
-}
-
-void PropertiesDock::updateActions()
-{
-    const QList<QtBrowserItem*> items = mPropertyBrowser->selectedItems();
-    bool allCustomProperties = !items.isEmpty() && mPropertyBrowser->allCustomPropertyItems(items);
-    bool editingTileset = mDocument && mDocument->type() == Document::TilesetDocumentType;
-    bool isTileset = mPropertyBrowser->object() && mPropertyBrowser->object()->isPartOfTileset();
-    bool canModify = allCustomProperties && (!isTileset || editingTileset);
-
-    // Disable remove and rename actions when none of the selected objects
-    // actually have the selected property (it may be inherited).
-    if (canModify) {
-        for (QtBrowserItem *item : items) {
-            if (!anyObjectHasProperty(mDocument->currentObjects(), item->property()->propertyName())) {
-                canModify = false;
-                break;
-            }
-        }
-    }
-
-    mActionRemoveProperty->setEnabled(canModify);
-    mActionRenameProperty->setEnabled(canModify && items.size() == 1);
-}
-
-void PropertiesDock::cutProperties()
-{
-    if (copyProperties())
-        removeProperties();
-}
-
-bool PropertiesDock::copyProperties()
-{
-    Object *object = mPropertyBrowser->object();
-    if (!object)
-        return false;
-
-    Properties properties;
-
-    const QList<QtBrowserItem*> items = mPropertyBrowser->selectedItems();
-    for (QtBrowserItem *item : items) {
-        if (!mPropertyBrowser->isCustomPropertyItem(item))
-            return false;
-
-        const QString name = item->property()->propertyName();
-        const QVariant value = object->property(name);
-        if (!value.isValid())
-            return false;
-
-        properties.insert(name, value);
-    }
-
-    ClipboardManager::instance()->setProperties(properties);
-    return true;
-}
-
-void PropertiesDock::pasteProperties()
-{
-    auto clipboardManager = ClipboardManager::instance();
-
-    Properties pastedProperties = clipboardManager->properties();
-    if (pastedProperties.isEmpty())
-        return;
-
-    const QList<Object *> objects = mDocument->currentObjects();
-    if (objects.isEmpty())
-        return;
-
-    QList<QUndoCommand*> commands;
-
-    for (Object *object : objects) {
-        Properties properties = object->properties();
-        mergeProperties(properties, pastedProperties);
-
-        if (object->properties() != properties) {
-            commands.append(new ChangeProperties(mDocument, QString(), object,
-                                                 properties));
-        }
-    }
-
-    if (!commands.isEmpty()) {
-        QUndoStack *undoStack = mDocument->undoStack();
-        undoStack->beginMacro(tr("Paste Property/Properties", nullptr,
-                                 pastedProperties.size()));
-
-        for (QUndoCommand *command : commands)
-            undoStack->push(command);
-
-        undoStack->endMacro();
-    }
-}
-
-void PropertiesDock::openAddPropertyDialog()
-{
-    AddPropertyDialog dialog(mPropertyBrowser);
-    if (dialog.exec() == AddPropertyDialog::Accepted)
-        addProperty(dialog.propertyName(), dialog.propertyValue());
-}
-
-void PropertiesDock::addProperty(const QString &name, const QVariant &value)
-{
-    if (name.isEmpty())
-        return;
-    Object *object = mDocument->currentObject();
-    if (!object)
-        return;
-
-    if (!object->hasProperty(name)) {
-        QUndoStack *undoStack = mDocument->undoStack();
-        undoStack->push(new SetProperty(mDocument,
-                                        mDocument->currentObjects(),
-                                        name, value));
-    }
-
-    mPropertyBrowser->editCustomProperty(name);
-}
-
-void PropertiesDock::removeProperties()
-{
-    Object *object = mDocument->currentObject();
-    if (!object)
-        return;
-
-    const QList<QtBrowserItem*> items = mPropertyBrowser->selectedItems();
-    if (items.isEmpty() || !mPropertyBrowser->allCustomPropertyItems(items))
-        return;
-
-    QStringList propertyNames;
-    for (QtBrowserItem *item : items)
-        propertyNames.append(item->property()->propertyName());
-
-    QUndoStack *undoStack = mDocument->undoStack();
-    undoStack->beginMacro(tr("Remove Property/Properties", nullptr,
-                             propertyNames.size()));
-
-    for (const QString &name : propertyNames) {
-        undoStack->push(new RemoveProperty(mDocument,
-                                           mDocument->currentObjects(),
-                                           name));
-    }
-
-    undoStack->endMacro();
-}
-
-void PropertiesDock::renameProperty()
-{
-    QtBrowserItem *item = mPropertyBrowser->currentItem();
-    if (!mPropertyBrowser->isCustomPropertyItem(item))
-        return;
-
-    const QString oldName = item->property()->propertyName();
-
-    QInputDialog *dialog = new QInputDialog(mPropertyBrowser);
-    dialog->setAttribute(Qt::WA_DeleteOnClose);
-    dialog->setInputMode(QInputDialog::TextInput);
-    dialog->setLabelText(tr("Name:"));
-    dialog->setTextValue(oldName);
-    dialog->setWindowTitle(tr("Rename Property"));
-    connect(dialog, &QInputDialog::textValueSelected, this, &PropertiesDock::renamePropertyTo);
-    dialog->open();
-}
-
-void PropertiesDock::renamePropertyTo(const QString &name)
-{
-    if (name.isEmpty())
-        return;
-
-    QtBrowserItem *item = mPropertyBrowser->currentItem();
-    if (!item)
-        return;
-
-    const QString oldName = item->property()->propertyName();
-    if (oldName == name)
-        return;
-
-    QUndoStack *undoStack = mDocument->undoStack();
-    undoStack->push(new RenameProperty(mDocument, mDocument->currentObjects(), oldName, name));
-}
-
-void PropertiesDock::showContextMenu(const QPoint &pos)
-{
-    const Object *object = mDocument->currentObject();
-    if (!object)
-        return;
-
-    const QList<QtBrowserItem *> items = mPropertyBrowser->selectedItems();
-    const bool customPropertiesSelected = !items.isEmpty() && mPropertyBrowser->allCustomPropertyItems(items);
-
-    bool currentObjectHasAllProperties = true;
-    QStringList propertyNames;
-    for (QtBrowserItem *item : items) {
-        const QString propertyName = item->property()->propertyName();
-        propertyNames.append(propertyName);
-
-        if (!object->hasProperty(propertyName))
-            currentObjectHasAllProperties = false;
-    }
-
-    QMenu contextMenu(mPropertyBrowser);
-
-    if (customPropertiesSelected && propertyNames.size() == 1) {
-        const auto value = object->resolvedProperty(propertyNames.first());
-        if (value.userType() == filePathTypeId()) {
-            const FilePath filePath = value.value<FilePath>();
-            const QString localFile = filePath.url.toLocalFile();
-
-            if (!localFile.isEmpty()) {
-                Utils::addOpenContainingFolderAction(contextMenu, localFile);
-
-                if (QFileInfo { localFile }.isFile())
-                    Utils::addOpenWithSystemEditorAction(contextMenu, localFile);
-            }
-        } else if (value.userType() == objectRefTypeId()) {
-            if (auto mapDocument = qobject_cast<MapDocument*>(mDocument)) {
-                const DisplayObjectRef objectRef(value.value<ObjectRef>(), mapDocument);
-
-                contextMenu.addAction(tr("Go to Object"), [=] {
-                    if (auto object = objectRef.object()) {
-                        objectRef.mapDocument->setSelectedObjects({object});
-                        emit objectRef.mapDocument->focusMapObjectRequested(object);
-                    }
-                })->setEnabled(objectRef.object());
-            }
-        }
-    }
-
-    if (!contextMenu.isEmpty())
-        contextMenu.addSeparator();
-
-    QAction *cutAction = contextMenu.addAction(tr("Cu&t"), this, &PropertiesDock::cutProperties);
-    QAction *copyAction = contextMenu.addAction(tr("&Copy"), this, &PropertiesDock::copyProperties);
-    QAction *pasteAction = contextMenu.addAction(tr("&Paste"), this, &PropertiesDock::pasteProperties);
-    contextMenu.addSeparator();
-    QMenu *convertMenu = nullptr;
-
-    if (customPropertiesSelected) {
-        convertMenu = contextMenu.addMenu(tr("Convert To"));
-        contextMenu.addAction(mActionRemoveProperty);
-        contextMenu.addAction(mActionRenameProperty);
-    } else {
-        contextMenu.addAction(mActionAddProperty);
-    }
-
-    cutAction->setShortcuts(QKeySequence::Cut);
-    cutAction->setIcon(QIcon(QLatin1String(":/images/16/edit-cut.png")));
-    cutAction->setEnabled(customPropertiesSelected && currentObjectHasAllProperties);
-    copyAction->setShortcuts(QKeySequence::Copy);
-    copyAction->setIcon(QIcon(QLatin1String(":/images/16/edit-copy.png")));
-    copyAction->setEnabled(customPropertiesSelected && currentObjectHasAllProperties);
-    pasteAction->setShortcuts(QKeySequence::Paste);
-    pasteAction->setIcon(QIcon(QLatin1String(":/images/16/edit-paste.png")));
-    pasteAction->setEnabled(ClipboardManager::instance()->hasProperties());
-
-    Utils::setThemeIcon(cutAction, "edit-cut");
-    Utils::setThemeIcon(copyAction, "edit-copy");
-    Utils::setThemeIcon(pasteAction, "edit-paste");
-
-    if (convertMenu) {
-        const int convertTo[] = {
-            QMetaType::Bool,
-            QMetaType::QColor,
-            QMetaType::Double,
-            filePathTypeId(),
-            objectRefTypeId(),
-            QMetaType::Int,
-            QMetaType::QString
-        };
-
-        // todo: could include custom property types
-
-        for (int toType : convertTo) {
-            bool someDifferentType = false;
-            bool allCanConvert = true;
-
-            for (const QString &propertyName : propertyNames) {
-                QVariant propertyValue = object->property(propertyName);
-
-                if (propertyValue.userType() != toType)
-                    someDifferentType = true;
-
-                if (!propertyValue.convert(toType)) {
-                    allCanConvert = false;
-                    break;
-                }
-            }
-
-            if (someDifferentType && allCanConvert) {
-                QAction *action = convertMenu->addAction(typeToName(toType));
-                action->setData(toType);
-            }
-        }
-
-        convertMenu->setEnabled(!convertMenu->actions().isEmpty());
-    }
-
-    ActionManager::applyMenuExtensions(&contextMenu, MenuIds::propertiesViewProperties);
-
-    const QPoint globalPos = mPropertyBrowser->mapToGlobal(pos);
-    const QAction *selectedItem = contextMenu.exec(globalPos);
-
-    if (selectedItem && convertMenu && selectedItem->parentWidget() == convertMenu) {
-        QUndoStack *undoStack = mDocument->undoStack();
-        undoStack->beginMacro(tr("Convert Property/Properties", nullptr, items.size()));
-
-        for (const QString &propertyName : propertyNames) {
-            QVariant propertyValue = object->property(propertyName);
-
-            int toType = selectedItem->data().toInt();
-            propertyValue.convert(toType);
-
-            undoStack->push(new SetProperty(mDocument,
-                                            mDocument->currentObjects(),
-                                            propertyName, propertyValue));
-        }
-
-        undoStack->endMacro();
-    }
+    mPropertiesWidget->selectCustomProperty(name);
 }
 
 bool PropertiesDock::event(QEvent *event)
 {
     switch (event->type()) {
-    case QEvent::ShortcutOverride: {
-        QKeyEvent *keyEvent = static_cast<QKeyEvent *>(event);
-        if (keyEvent->matches(QKeySequence::Delete) || keyEvent->key() == Qt::Key_Backspace
-                || keyEvent->matches(QKeySequence::Cut)
-                || keyEvent->matches(QKeySequence::Copy)
-                || keyEvent->matches(QKeySequence::Paste)) {
-            event->accept();
-            return true;
-        }
-        break;
-    }
     case QEvent::LanguageChange:
         retranslateUi();
         break;
@@ -505,32 +63,16 @@ bool PropertiesDock::event(QEvent *event)
     return QDockWidget::event(event);
 }
 
-void PropertiesDock::keyPressEvent(QKeyEvent *event)
+void PropertiesDock::bringToFront()
 {
-    if (event->matches(QKeySequence::Delete) || event->key() == Qt::Key_Backspace) {
-        removeProperties();
-    } else if (event->matches(QKeySequence::Cut)) {
-        cutProperties();
-    } else if (event->matches(QKeySequence::Copy)) {
-        copyProperties();
-    } else if (event->matches(QKeySequence::Paste)) {
-        pasteProperties();
-    } else {
-        QDockWidget::keyPressEvent(event);
-    }
+    show();
+    raise();
+    mPropertiesWidget->setFocus();
 }
 
 void PropertiesDock::retranslateUi()
 {
     setWindowTitle(tr("Properties"));
-
-    mActionAddProperty->setText(tr("Add Property"));
-
-    mActionRemoveProperty->setText(tr("Remove"));
-    mActionRemoveProperty->setToolTip(tr("Remove Property"));
-
-    mActionRenameProperty->setText(tr("Rename..."));
-    mActionRenameProperty->setToolTip(tr("Rename Property"));
 }
 
 } // namespace Tiled
