+++ src/tiled/propertieswidget.cpp
@@ -1,0 +1,521 @@
+/*
+ * propertieswidget.cpp
+ * Copyright 2013-2023, Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>
+ *
+ * This file is part of Tiled.
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; either version 2 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "propertieswidget.h"
+
+#include "actionmanager.h"
+#include "addpropertydialog.h"
+#include "changeproperties.h"
+#include "clipboardmanager.h"
+#include "mapdocument.h"
+#include "mapobject.h"
+#include "propertybrowser.h"
+#include "utils.h"
+#include "variantpropertymanager.h"
+
+#include <QAction>
+#include <QCoreApplication>
+#include <QEvent>
+#include <QFileInfo>
+#include <QInputDialog>
+#include <QKeyEvent>
+#include <QMenu>
+#include <QToolBar>
+#include <QUndoStack>
+#include <QVBoxLayout>
+
+namespace Tiled {
+
+PropertiesWidget::PropertiesWidget(QWidget *parent)
+    : QWidget{parent}
+    , mDocument(nullptr)
+    , mPropertyBrowser(new PropertyBrowser)
+{
+    mActionAddProperty = new QAction(this);
+    mActionAddProperty->setEnabled(false);
+    mActionAddProperty->setIcon(QIcon(QLatin1String(":/images/16/add.png")));
+    connect(mActionAddProperty, &QAction::triggered,
+            this, &PropertiesWidget::openAddPropertyDialog);
+
+    mActionRemoveProperty = new QAction(this);
+    mActionRemoveProperty->setEnabled(false);
+    mActionRemoveProperty->setIcon(QIcon(QLatin1String(":/images/16/remove.png")));
+    mActionRemoveProperty->setShortcuts(QKeySequence::Delete);
+    connect(mActionRemoveProperty, &QAction::triggered,
+            this, &PropertiesWidget::removeProperties);
+
+    mActionRenameProperty = new QAction(this);
+    mActionRenameProperty->setEnabled(false);
+    mActionRenameProperty->setIcon(QIcon(QLatin1String(":/images/16/rename.png")));
+    connect(mActionRenameProperty, &QAction::triggered,
+            this, &PropertiesWidget::renameProperty);
+
+    Utils::setThemeIcon(mActionAddProperty, "add");
+    Utils::setThemeIcon(mActionRemoveProperty, "remove");
+    Utils::setThemeIcon(mActionRenameProperty, "rename");
+
+    QToolBar *toolBar = new QToolBar;
+    toolBar->setFloatable(false);
+    toolBar->setMovable(false);
+    toolBar->setIconSize(Utils::smallIconSize());
+    toolBar->addAction(mActionAddProperty);
+    toolBar->addAction(mActionRemoveProperty);
+    toolBar->addAction(mActionRenameProperty);
+
+    QVBoxLayout *layout = new QVBoxLayout(this);
+    layout->setContentsMargins(0, 0, 0, 0);
+    layout->setSpacing(0);
+    layout->addWidget(mPropertyBrowser);
+    layout->addWidget(toolBar);
+    setLayout(layout);
+
+    mPropertyBrowser->setContextMenuPolicy(Qt::CustomContextMenu);
+    connect(mPropertyBrowser, &PropertyBrowser::customContextMenuRequested,
+            this, &PropertiesWidget::showContextMenu);
+    connect(mPropertyBrowser, &PropertyBrowser::selectedItemsChanged,
+            this, &PropertiesWidget::updateActions);
+
+    retranslateUi();
+}
+
+void PropertiesWidget::setDocument(Document *document)
+{
+    if (mDocument == document)
+        return;
+
+    if (mDocument)
+        mDocument->disconnect(this);
+
+    mDocument = document;
+    mPropertyBrowser->setDocument(document);
+
+    if (document) {
+        connect(document, &Document::currentObjectChanged,
+                this, &PropertiesWidget::currentObjectChanged);
+        connect(document, &Document::editCurrentObject,
+                this, &PropertiesWidget::bringToFront);
+
+        connect(document, &Document::propertyAdded,
+                this, &PropertiesWidget::updateActions);
+        connect(document, &Document::propertyRemoved,
+                this, &PropertiesWidget::updateActions);
+
+        currentObjectChanged(document->currentObject());
+    } else {
+        currentObjectChanged(nullptr);
+    }
+}
+
+void PropertiesWidget::selectCustomProperty(const QString &name)
+{
+    mPropertyBrowser->selectCustomProperty(name);
+}
+
+static bool anyObjectHasProperty(const QList<Object*> &objects, const QString &name)
+{
+    for (Object *obj : objects) {
+        if (obj->hasProperty(name))
+            return true;
+    }
+    return false;
+}
+
+void PropertiesWidget::currentObjectChanged(Object *object)
+{
+    mPropertyBrowser->setObject(object);
+
+    bool editingTileset = mDocument && mDocument->type() == Document::TilesetDocumentType;
+    bool isTileset = object && object->isPartOfTileset();
+    bool enabled = object && (!isTileset || editingTileset);
+
+    mPropertyBrowser->setEnabled(object);
+    mActionAddProperty->setEnabled(enabled);
+}
+
+void PropertiesWidget::updateActions()
+{
+    const QList<QtBrowserItem*> items = mPropertyBrowser->selectedItems();
+    bool allCustomProperties = !items.isEmpty() && mPropertyBrowser->allCustomPropertyItems(items);
+    bool editingTileset = mDocument && mDocument->type() == Document::TilesetDocumentType;
+    bool isTileset = mPropertyBrowser->object() && mPropertyBrowser->object()->isPartOfTileset();
+    bool canModify = allCustomProperties && (!isTileset || editingTileset);
+
+    // Disable remove and rename actions when none of the selected objects
+    // actually have the selected property (it may be inherited).
+    if (canModify) {
+        for (QtBrowserItem *item : items) {
+            if (!anyObjectHasProperty(mDocument->currentObjects(), item->property()->propertyName())) {
+                canModify = false;
+                break;
+            }
+        }
+    }
+
+    mActionRemoveProperty->setEnabled(canModify);
+    mActionRenameProperty->setEnabled(canModify && items.size() == 1);
+}
+
+void PropertiesWidget::cutProperties()
+{
+    if (copyProperties())
+        removeProperties();
+}
+
+bool PropertiesWidget::copyProperties()
+{
+    Object *object = mPropertyBrowser->object();
+    if (!object)
+        return false;
+
+    Properties properties;
+
+    const QList<QtBrowserItem*> items = mPropertyBrowser->selectedItems();
+    for (QtBrowserItem *item : items) {
+        if (!mPropertyBrowser->isCustomPropertyItem(item))
+            return false;
+
+        const QString name = item->property()->propertyName();
+        const QVariant value = object->property(name);
+        if (!value.isValid())
+            return false;
+
+        properties.insert(name, value);
+    }
+
+    ClipboardManager::instance()->setProperties(properties);
+    return true;
+}
+
+void PropertiesWidget::pasteProperties()
+{
+    auto clipboardManager = ClipboardManager::instance();
+
+    Properties pastedProperties = clipboardManager->properties();
+    if (pastedProperties.isEmpty())
+        return;
+
+    const QList<Object *> objects = mDocument->currentObjects();
+    if (objects.isEmpty())
+        return;
+
+    QList<QUndoCommand*> commands;
+
+    for (Object *object : objects) {
+        Properties properties = object->properties();
+        mergeProperties(properties, pastedProperties);
+
+        if (object->properties() != properties) {
+            commands.append(new ChangeProperties(mDocument, QString(), object,
+                                                 properties));
+        }
+    }
+
+    if (!commands.isEmpty()) {
+        QUndoStack *undoStack = mDocument->undoStack();
+        undoStack->beginMacro(tr("Paste Property/Properties", nullptr,
+                                 pastedProperties.size()));
+
+        for (QUndoCommand *command : commands)
+            undoStack->push(command);
+
+        undoStack->endMacro();
+    }
+}
+
+void PropertiesWidget::openAddPropertyDialog()
+{
+    AddPropertyDialog dialog(mPropertyBrowser);
+    if (dialog.exec() == AddPropertyDialog::Accepted)
+        addProperty(dialog.propertyName(), dialog.propertyValue());
+}
+
+void PropertiesWidget::addProperty(const QString &name, const QVariant &value)
+{
+    if (name.isEmpty())
+        return;
+    Object *object = mDocument->currentObject();
+    if (!object)
+        return;
+
+    if (!object->hasProperty(name)) {
+        QUndoStack *undoStack = mDocument->undoStack();
+        undoStack->push(new SetProperty(mDocument,
+                                        mDocument->currentObjects(),
+                                        name, value));
+    }
+
+    mPropertyBrowser->editCustomProperty(name);
+}
+
+void PropertiesWidget::removeProperties()
+{
+    Object *object = mDocument->currentObject();
+    if (!object)
+        return;
+
+    const QList<QtBrowserItem*> items = mPropertyBrowser->selectedItems();
+    if (items.isEmpty() || !mPropertyBrowser->allCustomPropertyItems(items))
+        return;
+
+    QStringList propertyNames;
+    for (QtBrowserItem *item : items)
+        propertyNames.append(item->property()->propertyName());
+
+    QUndoStack *undoStack = mDocument->undoStack();
+    undoStack->beginMacro(tr("Remove Property/Properties", nullptr,
+                             propertyNames.size()));
+
+    for (const QString &name : propertyNames) {
+        undoStack->push(new RemoveProperty(mDocument,
+                                           mDocument->currentObjects(),
+                                           name));
+    }
+
+    undoStack->endMacro();
+}
+
+void PropertiesWidget::renameProperty()
+{
+    QtBrowserItem *item = mPropertyBrowser->currentItem();
+    if (!mPropertyBrowser->isCustomPropertyItem(item))
+        return;
+
+    const QString oldName = item->property()->propertyName();
+
+    QInputDialog *dialog = new QInputDialog(mPropertyBrowser);
+    dialog->setAttribute(Qt::WA_DeleteOnClose);
+    dialog->setInputMode(QInputDialog::TextInput);
+    dialog->setLabelText(tr("Name:"));
+    dialog->setTextValue(oldName);
+    dialog->setWindowTitle(tr("Rename Property"));
+    connect(dialog, &QInputDialog::textValueSelected, this, &PropertiesWidget::renamePropertyTo);
+    dialog->open();
+}
+
+void PropertiesWidget::renamePropertyTo(const QString &name)
+{
+    if (name.isEmpty())
+        return;
+
+    QtBrowserItem *item = mPropertyBrowser->currentItem();
+    if (!item)
+        return;
+
+    const QString oldName = item->property()->propertyName();
+    if (oldName == name)
+        return;
+
+    QUndoStack *undoStack = mDocument->undoStack();
+    undoStack->push(new RenameProperty(mDocument, mDocument->currentObjects(), oldName, name));
+}
+
+void PropertiesWidget::showContextMenu(const QPoint &pos)
+{
+    const Object *object = mDocument->currentObject();
+    if (!object)
+        return;
+
+    const QList<QtBrowserItem *> items = mPropertyBrowser->selectedItems();
+    const bool customPropertiesSelected = !items.isEmpty() && mPropertyBrowser->allCustomPropertyItems(items);
+
+    bool currentObjectHasAllProperties = true;
+    QStringList propertyNames;
+    for (QtBrowserItem *item : items) {
+        const QString propertyName = item->property()->propertyName();
+        propertyNames.append(propertyName);
+
+        if (!object->hasProperty(propertyName))
+            currentObjectHasAllProperties = false;
+    }
+
+    QMenu contextMenu(mPropertyBrowser);
+
+    if (customPropertiesSelected && propertyNames.size() == 1) {
+        const auto value = object->resolvedProperty(propertyNames.first());
+        if (value.userType() == filePathTypeId()) {
+            const FilePath filePath = value.value<FilePath>();
+            const QString localFile = filePath.url.toLocalFile();
+
+            if (!localFile.isEmpty()) {
+                Utils::addOpenContainingFolderAction(contextMenu, localFile);
+
+                if (QFileInfo { localFile }.isFile())
+                    Utils::addOpenWithSystemEditorAction(contextMenu, localFile);
+            }
+        } else if (value.userType() == objectRefTypeId()) {
+            if (auto mapDocument = qobject_cast<MapDocument*>(mDocument)) {
+                const DisplayObjectRef objectRef(value.value<ObjectRef>(), mapDocument);
+
+                contextMenu.addAction(tr("Go to Object"), [=] {
+                    if (auto object = objectRef.object()) {
+                        objectRef.mapDocument->setSelectedObjects({object});
+                        emit objectRef.mapDocument->focusMapObjectRequested(object);
+                    }
+                })->setEnabled(objectRef.object());
+            }
+        }
+    }
+
+    if (!contextMenu.isEmpty())
+        contextMenu.addSeparator();
+
+    QAction *cutAction = contextMenu.addAction(tr("Cu&t"), this, &PropertiesWidget::cutProperties);
+    QAction *copyAction = contextMenu.addAction(tr("&Copy"), this, &PropertiesWidget::copyProperties);
+    QAction *pasteAction = contextMenu.addAction(tr("&Paste"), this, &PropertiesWidget::pasteProperties);
+    contextMenu.addSeparator();
+    QMenu *convertMenu = nullptr;
+
+    if (customPropertiesSelected) {
+        convertMenu = contextMenu.addMenu(tr("Convert To"));
+        contextMenu.addAction(mActionRemoveProperty);
+        contextMenu.addAction(mActionRenameProperty);
+    } else {
+        contextMenu.addAction(mActionAddProperty);
+    }
+
+    cutAction->setShortcuts(QKeySequence::Cut);
+    cutAction->setIcon(QIcon(QLatin1String(":/images/16/edit-cut.png")));
+    cutAction->setEnabled(customPropertiesSelected && currentObjectHasAllProperties);
+    copyAction->setShortcuts(QKeySequence::Copy);
+    copyAction->setIcon(QIcon(QLatin1String(":/images/16/edit-copy.png")));
+    copyAction->setEnabled(customPropertiesSelected && currentObjectHasAllProperties);
+    pasteAction->setShortcuts(QKeySequence::Paste);
+    pasteAction->setIcon(QIcon(QLatin1String(":/images/16/edit-paste.png")));
+    pasteAction->setEnabled(ClipboardManager::instance()->hasProperties());
+
+    Utils::setThemeIcon(cutAction, "edit-cut");
+    Utils::setThemeIcon(copyAction, "edit-copy");
+    Utils::setThemeIcon(pasteAction, "edit-paste");
+
+    if (convertMenu) {
+        const int convertTo[] = {
+            QMetaType::Bool,
+            QMetaType::QColor,
+            QMetaType::Double,
+            filePathTypeId(),
+            objectRefTypeId(),
+            QMetaType::Int,
+            QMetaType::QString
+        };
+
+        // todo: could include custom property types
+
+        for (int toType : convertTo) {
+            bool someDifferentType = false;
+            bool allCanConvert = true;
+
+            for (const QString &propertyName : propertyNames) {
+                QVariant propertyValue = object->property(propertyName);
+
+                if (propertyValue.userType() != toType)
+                    someDifferentType = true;
+
+                if (!propertyValue.convert(toType)) {
+                    allCanConvert = false;
+                    break;
+                }
+            }
+
+            if (someDifferentType && allCanConvert) {
+                QAction *action = convertMenu->addAction(typeToName(toType));
+                action->setData(toType);
+            }
+        }
+
+        convertMenu->setEnabled(!convertMenu->actions().isEmpty());
+    }
+
+    ActionManager::applyMenuExtensions(&contextMenu, MenuIds::propertiesViewProperties);
+
+    const QPoint globalPos = mPropertyBrowser->mapToGlobal(pos);
+    const QAction *selectedItem = contextMenu.exec(globalPos);
+
+    if (selectedItem && convertMenu && selectedItem->parentWidget() == convertMenu) {
+        QUndoStack *undoStack = mDocument->undoStack();
+        undoStack->beginMacro(tr("Convert Property/Properties", nullptr, items.size()));
+
+        for (const QString &propertyName : propertyNames) {
+            QVariant propertyValue = object->property(propertyName);
+
+            int toType = selectedItem->data().toInt();
+            propertyValue.convert(toType);
+
+            undoStack->push(new SetProperty(mDocument,
+                                            mDocument->currentObjects(),
+                                            propertyName, propertyValue));
+        }
+
+        undoStack->endMacro();
+    }
+}
+
+bool PropertiesWidget::event(QEvent *event)
+{
+    switch (event->type()) {
+    case QEvent::ShortcutOverride: {
+        QKeyEvent *keyEvent = static_cast<QKeyEvent *>(event);
+        if (keyEvent->matches(QKeySequence::Delete) || keyEvent->key() == Qt::Key_Backspace
+                || keyEvent->matches(QKeySequence::Cut)
+                || keyEvent->matches(QKeySequence::Copy)
+                || keyEvent->matches(QKeySequence::Paste)) {
+            event->accept();
+            return true;
+        }
+        break;
+    }
+    case QEvent::LanguageChange:
+        retranslateUi();
+        break;
+    default:
+        break;
+    }
+
+    return QWidget::event(event);
+}
+
+void PropertiesWidget::keyPressEvent(QKeyEvent *event)
+{
+    if (event->matches(QKeySequence::Delete) || event->key() == Qt::Key_Backspace) {
+        removeProperties();
+    } else if (event->matches(QKeySequence::Cut)) {
+        cutProperties();
+    } else if (event->matches(QKeySequence::Copy)) {
+        copyProperties();
+    } else if (event->matches(QKeySequence::Paste)) {
+        pasteProperties();
+    } else {
+        QWidget::keyPressEvent(event);
+    }
+}
+
+void PropertiesWidget::retranslateUi()
+{
+    mActionAddProperty->setText(QCoreApplication::translate("Tiled::PropertiesDock", "Add Property"));
+
+    mActionRemoveProperty->setText(QCoreApplication::translate("Tiled::PropertiesDock", "Remove"));
+    mActionRemoveProperty->setToolTip(QCoreApplication::translate("Tiled::PropertiesDock", "Remove Property"));
+
+    mActionRenameProperty->setText(QCoreApplication::translate("Tiled::PropertiesDock", "Rename..."));
+    mActionRenameProperty->setToolTip(QCoreApplication::translate("Tiled::PropertiesDock", "Rename Property"));
+}
+
+} // namespace Tiled
+
+#include "moc_propertieswidget.cpp"
